A fairly substantial recoding of the DSignJar class.
To allow for multiple files additional controls were needed. The
controls ensure the correct types of files are selected and to carefully
check signatures as well as duplicate file names. While this class
functions well on it's own the dialog will lock with multiple files
selected. This is a preliminary commit while threading is created to
alleviate a sizable multiple file selection as well as to enhance error
logging.