Updated
commented out the local sign jar however will keep it there to test for
the moment.

now clearing the folder text field when the more granular file input is
selected to better indicate the file was selected.

moved the prefix and suffix field check into the okpressed method as it
fits better with the other field checks conducted in the method.