JarInfoTableModel.load(): Remove Math.round()
The file.length() is an int and the 1024 is an int too and not a float like 1024.0.
So here we have an integer division rounded down.
But then we are converting the number to float and rounding it but there is nothing to round anymore.
So the Math.round() can be eliminated.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>