DCompareCertificates.initComponents():
The equals var is an int and the rows.size() is an int too and not a float.
So here we have an integer division rounded down.
But then we are converting the int to float.
Then the float is formatted but it remains rounded.
Instead we can just avoid the float conversion.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>