Custom Authority and Subject Key Identifiers (issue #418)
This PR harmonizes the naming of the public key finger print algorithms across dialogs, and users can now choose SHA-1 over SPKI or SHA-256 over SPKI when adding the Authority Key Identifier and Subject Key Identifier extensions.

This PR changes automatically populates the AKI for a signed certificate using the SKI of the issuing certificate. This helps ensure that certs signed by an issuer with a custom SKI are still identifiable when searching by the AKI.

This PR also fixes a bug that occurs when trying to edit an existing AKI containing a certificate serial number.

Known issues:

-   New tooltips are only in English
-   No warning displayed when generating a self signed cert with different methods for the AKI and SKI.